+++ scenedetect/output/image.py
@@ -293,7 +293,8 @@ class _ImageExtractor:
     def generate_timecode_list(self, scene_list: SceneList) -> ty.List[ty.Iterable[FrameTimecode]]:
         """Generates a list of timecodes for each scene in `scene_list` based on the current config
         parameters."""
-        framerate = scene_list[0][0]._framerate
+        # TODO(v0.7): This needs to be fixed as part of PTS overhaul.
+        framerate = scene_list[0][0].framerate
         # TODO(v1.0): Split up into multiple sub-expressions so auto-formatter works correctly.
         return [
             (
@@ -450,7 +451,7 @@ def save_images(
     image_num_format = "%0"
     image_num_format += str(math.floor(math.log(num_images, 10)) + 2) + "d"
 
-    framerate = scene_list[0][0]._framerate
+    framerate = scene_list[0][0]._rate
 
     # TODO(v1.0): Split up into multiple sub-expressions so auto-formatter works correctly.
     timecode_list = [